Implement Conversations feature and enhance App routing
- Added a new Conversations component to display past conversations with options to delete and load more.
- Updated App component to include a route for the Conversations page, ensuring proper navigation based on authentication status.
- Enhanced Layout component to accommodate the new Conversations section in the sidebar.
- Introduced new styles for the Conversations page to improve user experience and visual appeal.
- Refactored ChatInterface to remove unused conversation handling logic, streamlining the component's functionality.
