+++ apps/client/src/components/ChatInterface.tsx
@@ -3,10 +3,9 @@ import ReactMarkdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
 import { useParams, useNavigate } from 'react-router-dom';
 import { apiService } from '../services/apiService';
-import type { 
-  ChatMessage, 
-  ChatRequest, 
-  Conversation,
+import type {
+  ChatMessage,
+  ChatRequest,
   Provider
 } from '@kronos/core';
 import { StreamEvent, StreamEventType } from '@kronos/core';
@@ -21,7 +20,7 @@ interface ChatInterfaceProps {
 const ChatInterface: React.FC<ChatInterfaceProps> = () => {
   const { conversationId: urlConversationId } = useParams<{ conversationId?: string }>();
   const navigate = useNavigate();
-  
+
   const [messages, setMessages] = useState<ChatMessage[]>([]);
   const [input, setInput] = useState('');
   const [isStreaming, setIsStreaming] = useState(false);
@@ -30,22 +29,12 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
   const [error, setError] = useState<string | null>(null);
   const [streamingMessage, setStreamingMessage] = useState('');
   const [streamingMarkdown, setStreamingMarkdown] = useState('');
-  const [conversations, setConversations] = useState<Conversation[]>([]);
-  const [showConversations, setShowConversations] = useState(false);
-  const [isLoadingConversations, setIsLoadingConversations] = useState(false);
-  const [hasMoreConversations, setHasMoreConversations] = useState(true);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [totalConversations, setTotalConversations] = useState(0);
-  const [showDeleteConfirmation, setShowDeleteConfirmation] = useState(false);
-  const [conversationToDelete, setConversationToDelete] = useState<Conversation | null>(null);
-  const [isDeletingConversation, setIsDeletingConversation] = useState(false);
   const [progressUpdate, setProgressUpdate] = useState<string | null>(null);
   const [selectedToolkits, setSelectedToolkits] = useState<Provider[]>([]);
-  
+
   const messagesEndRef = useRef<HTMLDivElement>(null);
   const textareaRef = useRef<HTMLTextAreaElement>(null);
   const abortControllerRef = useRef<AbortController | null>(null);
-  const conversationsListRef = useRef<HTMLDivElement>(null);
 
   // Generate a title from the first message
   const generateTitleFromMessage = (message: string): string => {
@@ -95,47 +84,6 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
     }
   }, [urlConversationId]);
 
-  // Handle escape key to close modals
-  useEffect(() => {
-    const handleEscape = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') {
-        if (showDeleteConfirmation) {
-          cancelDeleteConversation();
-        } else if (showConversations) {
-          setShowConversations(false);
-        }
-      }
-    };
-
-    if (showConversations || showDeleteConfirmation) {
-      document.addEventListener('keydown', handleEscape);
-      return () => document.removeEventListener('keydown', handleEscape);
-    }
-    
-    return undefined;
-  }, [showConversations, showDeleteConfirmation]);
-
-  // Handle infinite scroll for conversations
-  useEffect(() => {
-    const handleScroll = () => {
-      if (!conversationsListRef.current || !showConversations) return;
-      
-      const { scrollTop, scrollHeight, clientHeight } = conversationsListRef.current;
-      const isNearBottom = scrollTop + clientHeight >= scrollHeight - 100; // 100px threshold
-      
-      if (isNearBottom && hasMoreConversations && !isLoadingConversations) {
-        loadMoreConversations();
-      }
-    };
-
-    const listElement = conversationsListRef.current;
-    if (listElement) {
-      listElement.addEventListener('scroll', handleScroll);
-      return () => listElement.removeEventListener('scroll', handleScroll);
-    }
-    
-    return undefined;
-  }, [showConversations, hasMoreConversations, isLoadingConversations, currentPage]);
 
   // Persist conversation ID to localStorage when it changes
   useEffect(() => {
@@ -147,103 +95,6 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
   }, [currentConversationId]);
 
 
-  const loadConversations = async (page: number = 1, append: boolean = false) => {
-    if (isLoadingConversations) return;
-    
-    setIsLoadingConversations(true);
-    try {
-      const response = await apiService.getConversations(page, 10);
-      
-      if (append) {
-        setConversations(prev => [...prev, ...response.items]);
-      } else {
-        setConversations(response.items || []);
-      }
-      
-      // Use the actual total from API response, not calculated
-      setTotalConversations(response.total);
-      setHasMoreConversations(page < response.totalPages);
-      setCurrentPage(page);
-      setError(null); // Clear any previous errors
-    } catch (error: any) {
-      console.error('Failed to load conversations:', error);
-      
-      // Handle different types of errors
-      if (error.response?.status === 400) {
-        setError(`Invalid request: ${error.response.data.message || 'Please check your parameters'}`);
-      } else if (error.response?.status === 401) {
-        setError('Authentication required. Please log in again.');
-      } else if (error.response?.status === 500) {
-        setError('Server error. Please try again later.');
-      } else if (error.code === 'NETWORK_ERROR') {
-        setError('Network error. Please check your connection.');
-      } else {
-        setError('Failed to load conversations. Please try again.');
-      }
-    } finally {
-      setIsLoadingConversations(false);
-    }
-  };
-
-  const loadMoreConversations = async () => {
-    if (hasMoreConversations && !isLoadingConversations) {
-      await loadConversations(currentPage + 1, true);
-    }
-  };
-
-  const showDeleteConfirmationModal = (conversation: Conversation, event: React.MouseEvent) => {
-    event.stopPropagation(); // Prevent triggering the conversation load
-    setConversationToDelete(conversation);
-    setShowDeleteConfirmation(true);
-  };
-
-  const confirmDeleteConversation = async () => {
-    if (!conversationToDelete) return;
-
-    setIsDeletingConversation(true);
-    try {
-      const result = await apiService.deleteConversation(conversationToDelete.id);
-      if (result.success) {
-        // Remove from local state
-        setConversations(prev => prev.filter(conv => conv.id !== conversationToDelete.id));
-        
-        // If this was the current conversation, clear it
-        if (currentConversationId === conversationToDelete.id) {
-          startNewConversation();
-        }
-        setError(null); // Clear any previous errors
-      } else {
-        setError(result.message || 'Failed to delete conversation');
-      }
-    } catch (error: any) {
-      console.error('Failed to delete conversation:', error);
-      
-      // Handle different types of errors
-      if (error.response?.status === 400) {
-        setError(`Invalid request: ${error.response.data.message || 'Please check your parameters'}`);
-      } else if (error.response?.status === 401) {
-        setError('Authentication required. Please log in again.');
-      } else if (error.response?.status === 404) {
-        setError('Conversation not found.');
-      } else if (error.response?.status === 500) {
-        setError('Server error. Please try again later.');
-      } else if (error.code === 'NETWORK_ERROR') {
-        setError('Network error. Please check your connection.');
-      } else {
-        setError('Failed to delete conversation. Please try again.');
-      }
-    } finally {
-      setIsDeletingConversation(false);
-      setShowDeleteConfirmation(false);
-      setConversationToDelete(null);
-    }
-  };
-
-  const cancelDeleteConversation = () => {
-    setShowDeleteConfirmation(false);
-    setConversationToDelete(null);
-  };
-
   const loadConversationMessages = async (conversationId: string) => {
     try {
       const response = await apiService.getConversationMessages(conversationId);
@@ -252,20 +103,17 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
       setStreamingMessage('');
       setStreamingMarkdown('');
       setError(null);
-      
+
       // Update URL if not already there
       if (urlConversationId !== conversationId) {
         navigate(`/chat/${conversationId}`, { replace: true });
       }
-      
+
       // Set conversation title from the conversation data
-      const conversation = conversations.find(c => c.id === conversationId);
-      if (conversation) {
-        setConversationTitle(conversation.title || 'Untitled Conversation');
-      }
+      setConversationTitle('Untitled Conversation');
     } catch (error: any) {
       console.error('Failed to load conversation messages:', error);
-      
+
       // Handle different types of errors
       if (error.response?.status === 400) {
         setError(`Invalid request: ${error.response.data.message || 'Please check your parameters'}`);
@@ -289,10 +137,6 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
     setConversationTitle('');
     setStreamingMessage('');
     setError(null);
-    setShowConversations(false);
-    // Reset pagination state
-    setCurrentPage(1);
-    setHasMoreConversations(true);
     // Navigate to base chat URL
     navigate('/chat', { replace: true });
   };
@@ -356,7 +200,7 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
       try {
         while (true) {
           const { done, value } = await reader.read();
-          
+
           if (done) break;
 
           const chunk = decoder.decode(value, { stream: true });
@@ -365,39 +209,39 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
           for (const line of lines) {
             if (line.startsWith('data: ')) {
               const data = line.slice(6);
-              
+
               if (data === '[DONE]') {
                 break;
               }
 
               try {
                 const parsed: StreamEvent = JSON.parse(data);
-                
+
                 switch (parsed.type) {
                   case StreamEventType.START:
                     conversationId = (parsed.data as any).conversationId || '';
                     sessionId = (parsed.data as any).sessionId || '';
                     isNewConversation = (parsed.data as any).isNewConversation || false;
                     setCurrentConversationId(conversationId);
-                    
+
                     // Update URL with new conversation ID
                     if (conversationId && urlConversationId !== conversationId) {
                       navigate(`/chat/${conversationId}`, { replace: true });
                     }
-                    
+
                     console.log('Stream started:', { conversationId, sessionId, isNewConversation });
                     break;
-                    
+
                   case StreamEventType.TOKEN:
                     assistantMessage += (parsed.data as any).token || '';
                     setStreamingMessage(assistantMessage);
                     setStreamingMarkdown(assistantMessage);
                     break;
-                    
+
                   case StreamEventType.PROGRESS_UPDATE:
                     setProgressUpdate((parsed.data as any).message || '');
                     break;
-                    
+
                   case StreamEventType.END: {
                     // Finalize the assistant message
                     const finalMessage: ChatMessage = {
@@ -409,14 +253,14 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
                     setStreamingMessage('');
                     setProgressUpdate(null);
                     setIsStreaming(false);
-                    console.log('Stream ended:', { 
-                      conversationId, 
+                    console.log('Stream ended:', {
+                      conversationId,
                       totalTokens: parsed.data.totalTokens,
-                      processingTime: parsed.data.processingTime 
+                      processingTime: parsed.data.processingTime
                     });
                     return;
                   }
-                    
+
                   default:
                     console.warn('Unknown stream event type:', parsed.type);
                 }
@@ -461,9 +305,9 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
 
 
   const formatTimestamp = (timestamp: string) => {
-    return new Date(timestamp).toLocaleTimeString([], { 
-      hour: '2-digit', 
-      minute: '2-digit' 
+    return new Date(timestamp).toLocaleTimeString([], {
+      hour: '2-digit',
+      minute: '2-digit'
     });
   };
 
@@ -473,24 +317,9 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
       <div className="chat-header-section">
         <div className="chat-header">
           <div className="chat-header-left">
-            <button
-              onClick={() => {
-                if (!showConversations) {
-                  // Always load conversations when opening the modal
-                  loadConversations();
-                }
-                setShowConversations(!showConversations);
-              }}
-              className="chat-control-btn"
-              title="Conversations"
-            >
-              <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
-              </svg>
-            </button>
             <p className="text-sm text-gray-300">
-              {currentConversationId 
-                ? (conversationTitle || conversations.find(c => c.id === currentConversationId)?.title || 'Untitled Conversation')
+              {currentConversationId
+                ? (conversationTitle || 'Untitled Conversation')
                 : (conversationTitle || 'New conversation')
               }
             </p>
@@ -510,176 +339,6 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
         </div>
       </div>
 
-      {/* Past Conversations Section */}
-      {showConversations && (
-        <div className="conversations-section">
-          <div 
-            className="conversations-modal-overlay"
-            onClick={(e) => {
-              if (e.target === e.currentTarget) {
-                setShowConversations(false);
-              }
-            }}
-          >
-          <div className="conversations-modal">
-            <div className="conversations-modal-header">
-              <div className="conversations-header-content">
-                <h3>Past Conversations</h3>
-                <p className="conversations-subtitle">
-                  {totalConversations === 0 
-                    ? '0 Conversations found'
-                    : `${conversations.length} of ${totalConversations} conversation${totalConversations !== 1 ? 's' : ''} loaded`
-                  }
-                </p>
-              </div>
-              <button
-                onClick={() => setShowConversations(false)}
-                className="conversations-modal-close"
-                aria-label="Close conversations"
-              >
-                <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                </svg>
-              </button>
-            </div>
-            <div className="conversations-modal-content">
-              <div className="conversations-list" ref={conversationsListRef}>
-                {/* Conversations List */}
-                {conversations.length === 0 && !isLoadingConversations ? (
-                  <div className="conversations-empty">
-                    <div className="conversations-empty-icon">
-                      <svg className="w-16 h-16 text-gray-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z" />
-                      </svg>
-                    </div>
-                    <h4 className="conversations-empty-title">No conversations yet</h4>
-                    <p className="conversations-empty-description">
-                      Start chatting to see your conversation history here
-                    </p>
-                  </div>
-                ) : (
-                  conversations.map((conversation) => (
-                    <div
-                      key={conversation.id}
-                      className={`conversation-item ${currentConversationId === conversation.id ? 'active' : ''}`}
-                    >
-                      <button
-                        onClick={() => {
-                          loadConversationMessages(conversation.id);
-                          setShowConversations(false);
-                        }}
-                        className="conversation-content"
-                      >
-                        <div className="conversation-icon">
-                          <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                            <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M8 12h.01M12 12h.01M16 12h.01M21 12c0 4.418-4.03 8-9 8a9.863 9.863 0 01-4.255-.949L3 20l1.395-3.72C3.512 15.042 3 13.574 3 12c0-4.418 4.03-8 9-8s9 3.582 9 8z" />
-                          </svg>
-                        </div>
-                        <div className="conversation-details">
-                          <div className="conversation-title">
-                            {conversation.title || 'Untitled Conversation'}
-                          </div>
-                          <div className="conversation-time">
-                            {new Date(conversation.updatedAt).toLocaleDateString('en-US', {
-                              month: 'short',
-                              day: 'numeric',
-                              year: 'numeric',
-                              hour: '2-digit',
-                              minute: '2-digit'
-                            })}
-                          </div>
-                        </div>
-                      </button>
-                      <button
-                        onClick={(e) => showDeleteConfirmationModal(conversation, e)}
-                        className="conversation-delete-btn"
-                        title="Delete conversation"
-                        aria-label="Delete conversation"
-                      >
-                        <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
-                        </svg>
-                      </button>
-                    </div>
-                  ))
-                )}
-                
-                {/* Loading indicator */}
-                {isLoadingConversations && (
-                  <div className="conversations-loading">
-                    <div className="conversations-loading-spinner"></div>
-                    <span>Loading more conversations...</span>
-                  </div>
-                )}
-                
-                {/* End of list indicator */}
-                {!hasMoreConversations && conversations.length > 0 && (
-                  <div className="conversations-end">
-                    <div className="conversations-end-line"></div>
-                    <span>You've reached the end</span>
-                    <div className="conversations-end-line"></div>
-                  </div>
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-        </div>
-      )}
-
-      {/* Delete Confirmation Modal */}
-      {showDeleteConfirmation && conversationToDelete && (
-        <div 
-          className="delete-confirmation-overlay"
-          onClick={(e) => {
-            if (e.target === e.currentTarget) {
-              cancelDeleteConversation();
-            }
-          }}
-        >
-          <div className="delete-confirmation-modal">
-            <div className="delete-confirmation-header">
-              <h3>Delete Conversation</h3>
-            </div>
-            <div className="delete-confirmation-content">
-              <p>Are you sure you want to delete this conversation?</p>
-              <div className="conversation-preview">
-                <strong>{conversationToDelete.title || `Conversation ${conversationToDelete.id.slice(-8)}`}</strong>
-                <span className="conversation-date">
-                  {new Date(conversationToDelete.updatedAt).toLocaleDateString()}
-                </span>
-              </div>
-              <p className="warning-text">This action cannot be undone.</p>
-            </div>
-            <div className="delete-confirmation-actions">
-              <button
-                onClick={cancelDeleteConversation}
-                className="delete-confirmation-btn cancel"
-                disabled={isDeletingConversation}
-              >
-                Cancel
-              </button>
-              <button
-                onClick={confirmDeleteConversation}
-                className="delete-confirmation-btn delete"
-                disabled={isDeletingConversation}
-              >
-                {isDeletingConversation ? (
-                  <>
-                    <svg className="animate-spin h-4 w-4 mr-2" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                      <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                      <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                    </svg>
-                    Deleting...
-                  </>
-                ) : (
-                  'Delete'
-                )}
-              </button>
-            </div>
-          </div>
-        </div>
-      )}
 
       {/* Messages Area */}
       <div className="chat-messages">
@@ -700,7 +359,7 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
             <div className={`message-bubble ${message.role}`}>
               <div className="whitespace-pre-wrap break-words">
                 {message.role === 'assistant' ? (
-                  <ReactMarkdown 
+                  <ReactMarkdown
                     remarkPlugins={[remarkGfm]}
                     components={{
                       code: ({ node, className, children, ...props }: any) => {
@@ -773,7 +432,7 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
           <div className="chat-message assistant">
             <div className="message-bubble assistant">
               <div className="whitespace-pre-wrap break-words">
-                <ReactMarkdown 
+                <ReactMarkdown
                   remarkPlugins={[remarkGfm]}
                   components={{
                     code: ({ node, className, children, ...props }: any) => {
@@ -854,7 +513,7 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
               onToolkitsChange={setSelectedToolkits}
             />
           </div>
-          
+
           <div className="chat-input-main">
             <textarea
               ref={textareaRef}
@@ -872,16 +531,16 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
               className={`chat-send-btn ${isStreaming ? 'streaming' : ''}`}
             >
               {isStreaming ? (
-                <img 
-                  src="/images/integrations/stop.png" 
-                  alt="Stop" 
+                <img
+                  src="/images/integrations/stop.png"
+                  alt="Stop"
                   className="h-5 w-5"
                   style={{ objectFit: 'contain' }}
                 />
               ) : (
-                <img 
-                  src="/images/integrations/send.png" 
-                  alt="Send" 
+                <img
+                  src="/images/integrations/send.png"
+                  alt="Send"
                   className="h-4 w-4"
                   style={{ objectFit: 'contain' }}
                 />
@@ -889,7 +548,7 @@ const ChatInterface: React.FC<ChatInterfaceProps> = () => {
             </button>
           </div>
         </div>
-        
+
         <div className="chat-disclaimer">
           Kronos AI can make mistakes. Consider checking important information.
         </div>
