+++ apps/client/src/components/Sidebar.tsx
@@ -31,6 +31,16 @@ const Sidebar: React.FC<SidebarProps> = ({
       )
     },
     {
+      id: 'conversations',
+      label: 'Conversations',
+      path: '/conversations',
+      icon: (
+        <svg className="w-5 h-5" fill="currentColor" viewBox="0 0 20 20">
+          <path fillRule="evenodd" d="M3 4a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zm0 4a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zm0 4a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1zm0 4a1 1 0 011-1h12a1 1 0 110 2H4a1 1 0 01-1-1z" clipRule="evenodd" />
+        </svg>
+      )
+    },
+    {
       id: 'integrations',
       label: 'Integrations',
       path: '/oauth-integrations',
